Split report generation into content and renderer layers
